add regions poller
