+++ ec2list.go
@@ -1,0 +1,159 @@
+package main
+
+import (
+	"fmt"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/awserr"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/ec2"
+	"log"
+	"strings"
+)
+
+var (
+	// @readonly
+	defaultRegion        = "us-west-1"
+	awsErrError          = "[ERROR] %+v %+v %+v"
+	awsErrRequestFailure = "[ERROR] %+v %+v %+v %+v"
+)
+
+type EC2List struct {
+	EC2     *ec2.EC2
+	Profile *Profile
+	Region  string
+}
+
+// Returns a pointer to a new EC2List object
+func NewEC2List(profile *Profile, region string) *EC2List {
+	// If region is not specified, connect to default one - us-west-1
+	if region == "" {
+		region = defaultRegion
+	}
+
+	config := aws.Config{
+		Region:      aws.String(region),
+		Credentials: profile.Credentials,
+	}
+
+	return &EC2List{
+		EC2:     ec2.New(session.New(), &config),
+		Profile: profile,
+		Region:  region,
+	}
+}
+
+// ListInstances print list of instances in a format:
+// {id},{name},{private_ip},{instance_size},{public_ip},{region},{account}
+func (c *EC2List) ListInstances(token string) {
+	defer wg.Done()
+	// Prepare request
+	params := &ec2.DescribeInstancesInput{
+		DryRun: aws.Bool(false),
+		Filters: []*ec2.Filter{
+			{
+				// Return only "running" and "pending" instances
+				Name: aws.String("instance-state-name"),
+				Values: []*string{
+					aws.String("running"),
+					aws.String("pending"),
+				},
+			},
+		},
+		// Maximum count instances on one result page
+		MaxResults: aws.Int64(1000),
+		// Next page token
+		NextToken: aws.String(token),
+	}
+
+	// Get list of ec2 instances
+	res, err := c.EC2.DescribeInstances(params)
+	if err != nil {
+		if awsErr, ok := err.(awserr.Error); ok {
+			log.Printf(awsErrError, awsErr.Code(), awsErr.Message(), awsErr.OrigErr())
+			if reqErr, ok := err.(awserr.RequestFailure); ok {
+				log.Printf(awsErrRequestFailure, reqErr.Code(), reqErr.Message(), reqErr.StatusCode(), reqErr.RequestID())
+			}
+		} else {
+			log.Printf(err.Error())
+		}
+	}
+
+	// Extract instances info from result and print it
+	for _, reservation := range res.Reservations {
+		for _, instance := range reservation.Instances {
+
+			// If there is no tag "Name", return "None"
+			name := "None"
+			for _, keys := range instance.Tags {
+				if *keys.Key == "Name" {
+					name = *keys.Value
+				}
+			}
+
+			instance_string := []*string{
+				instance.InstanceId,
+				&name,
+				instance.PrivateIpAddress,
+				instance.InstanceType,
+				instance.PublicIpAddress,
+				&c.Region,
+				&c.Profile.Name,
+			}
+
+			output_string := []string{}
+			for _, str := range instance_string {
+				if str == nil {
+					output_string = append(output_string, "None")
+				} else {
+					output_string = append(output_string, *str)
+				}
+			}
+
+			instance := strings.Join(output_string, ",")
+			// If running in service mode, write in output buffer, else just print
+			if *service {
+				output_buffer = append(output_buffer, instance)
+			} else {
+				fmt.Printf("%s\n", instance)
+			}
+		}
+	}
+
+	// If there are more instances repeat request with a token
+	if res.NextToken != nil {
+		wg.Add(1)
+		go c.ListInstances(*res.NextToken)
+	}
+}
+
+// Run go routines to print all instances from all regions and all accounts
+func fetchInstances() {
+	var profile *Profile
+
+	// Clear output_buffer
+	output_buffer = []string{}
+
+	// Run go routines to print instances
+	for _, profile_name := range profiles {
+		// If we didn't load regions already, then fill regions slice
+		wg.Add(1)
+		profile = NewProfile(profile_name)
+		go PrintInstances(profile, regions)
+	}
+
+	// Wait until receive info about all instances
+	wg.Wait()
+
+	// Resize and fill screen buffer with output data
+	screen_buffer = make([]string, len(output_buffer), (cap(output_buffer)+1)*2)
+	copy(screen_buffer, output_buffer)
+}
+
+// PrintInstances runs go routines to print instances from all regions within account
+func PrintInstances(profile *Profile, regions []string) {
+	defer wg.Done()
+	for _, region := range regions {
+		wg.Add(1)
+		go NewEC2List(profile, region).ListInstances("")
+	}
+}
